e145/sheldonbrain-omega/core/grokbrain_v4/requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.12.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.0.1
sentence-transformers>=2.2.2
qdrant-client>=1.7.0
structlog>=23.2.0
streamlit>=1.30.0
pandas>=2.0.0
transformers>=4.36.0
torch>=2.1.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.108.0
